+++ src/window.rs
@@ -9,8 +9,11 @@ use std::{
     os::raw,
 };
 
+/// Opaque raw window handle (*mut c_void to HWND on Windows and NSWindow on MacOS)
 #[cfg(any(target_os = "windows", target_os = "macos", target_os = "ios", target_os = "android"))]
 pub type RawHandle = *mut raw::c_void;
+
+/// XID (u64) raw window handle for X11
 #[cfg(any(
     target_os = "linux",
     target_os = "dragonfly",